second try of mainService of graduated
a huge number of corrections of remarks in accordance with the review

+++ stats-service/dto/src/main/java/ru/practicum/dto/ViewStatsDto.java
@@ -14,5 +14,5 @@ public class ViewStatsDto {
 
     String uri;
 
-    Integer hits;
+    Long hits;
 }
